Move VariantTensorDataWriter and VariantTensorDataReader to dataset_utils
